Added 'SellSystem'
Creates a thread that looks for all assets that I won in  every single wallet that belongs to my account and gets the ticker symbol, the sale price and the fees used to buy the asset, if the current price of the asset plus the sale fees is more than the sale price and buy fees, sell the asset for profit.

If the asset doesn't meet any of the conditions, keep it till the price change of the asset in the last 24 hours increases and revisit.